+++ pallets/reversible-transfers/src/tests/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+mod mock;
+
+#[cfg(test)]
+mod tests;
+
+#[cfg(test)]
+mod test_high_security_account;
